concept-api: Fix linting errors after adding `sbt-tpolecat`

+++ concept-api/src/main/scala/no/ndla/conceptapi/model/api/TagsSearchResult.scala
@@ -13,7 +13,7 @@ import scala.annotation.meta.field
 
 @ApiModel(description = "Information about tags-search-results")
 case class TagsSearchResult(
-    @(ApiModelProperty @field)(description = "The total number of tags matching this query") totalCount: Long,
+    @(ApiModelProperty @field)(description = "The total number of tags matching this query") totalCount: Int,
     @(ApiModelProperty @field)(description = "For which page results are shown from") page: Int,
     @(ApiModelProperty @field)(description = "The number of results per page") pageSize: Int,
     @(ApiModelProperty @field)(description = "The chosen search language") language: String,
